Spawn new cell at available positions on the world
Test that two cells cannot overlap.
Test that the world retrieves a correct list of cells when asked to get either a specific row or column.
Test that the world cannot creates a new cell when there is no more space left.
And as always, a bit of refactor (could be better though...).