Merging neighbours cells of same value of a list
The Cell's sorting method is now a static method of the GameLogic.
The Enum CellComparator changed a bit. Now there is four types of it:
(ROW_ASC, ROW_DESC, COLUMN_ASC, COLUMN_DESC).
Even if there is a duplicate in the code, this is now more clear/readable.
Adding of a new test file for the game logic.
And a bit of refactoring.